backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
web3==6.20.3
apscheduler==3.10.4
psycopg2-binary==2.9.10
httpx==0.27.2
python-dotenv==1.0.1
numpy==2.1.3
pandas==2.2.3
scipy==1.13.1
scikit-learn==1.5.2
dbt-core==1.8.7
dbt-postgres==1.8.2
python-dateutil==2.9.0.post0
